Penalize spoofing of some Brazilian domains
There was a recent data breach that revealed sensitive information for
220 million Brazilians [1]: names, addresses, social security numbers,
medical records, employment information, and more.

Due to this information now becoming public, phishing e-mails became way
more effective since they can be personalized and contain accurate
information about the recipient, giving the impression of being a legit
message sent by a bank or the government [2].

Therefore it becomes quite important that e-mail which is spoofing
legitimate senders is mistakenly classified as spam. This commit ensures
that if DMARC validation fails for such critical domains, the e-mail
is penalized accordingly.

[1] https://www.schneier.com/blog/archives/2021/01/massive-brazilian-data-breach.html
[2] The author of this commit was a recipient of such messages.